Add "Build Duration" time to the published AMQP event
This change will help Jenkins operation teams to have
a better understanding around the duration of any
specific jobs when they are completed. They can directly
relate this information to cost when the underlying
executor resources are billed by time.
e.g. hourly rate applied to slaves based on-demand AWS EC2

At SoMC, we intend to ship and index all the events coming
into the configured RabbitMQ exchange to our Elasticsearch
server so that we can analyze and visualize the data to
improve our daily operation.