Fix message formatting problem for multi-line parameters
If a queued item contained parameters whit values spanning over
multiple lines, then the parameter list in the resulting message was
badly formatted. E.g the parameter FOO='BAR\nBAZ' was encoded as:

parameters:["FOO=BAR","BAZ"]

This change fixes that problem as well as unifying and improving
handling of parameters for queued items and started / completed builds.
This is done by ensuring that it is the same code which produces the
parameters list for all events types.