Fixed findbugs warnings
Took care of some findbugs warnings that were added
years ago.
Findbugs was also upgraded due to a bug in 3.0.1.

The main findbugs fix that has been done is that the initialization
of messageQueue and messageQueueThread are now synchronized, in
MQConnection.addMessageToQueue
Since every QueueListenerImpl and RunListenerImpl could enter the method
simultaneously this was needed.