Don't start message queue thread until after initialization
For some time there's been a race condition during Jenkins startup
between the sending of the first event and the initialization of
the message queue thread.

The AMQP configuration is passed to MQConnection via initialize(),
called from MQNotifierConfig#configure(). Until this has happened,
the MQConnection's AMQP URL will be null and any attempt to send an
event will trigger an NPE.

We solve this by introducing an MQConnection#initialized attribute
that's set to true only after we've received the configuration, and
until then the message queue thread won't be started. We still post
messages to the queue, though, so any early messages will eventually
be sent.