More fine grained routing key settings for messages
Currently, the only way to choose a routing key is to set one
which will be used for every message.
This commit adds an automatic setting of routing keys, where
each method that publishes messages gets one based on the method.
For details on the automatic routing keys, look at:
src/main/webapp/help-automatic-routing-key.html